+++ app/(authorized)/profile/[id]/[userName]/page.tsx
@@ -1,4 +1,4 @@
-import {UserTabs} from "@/widget/userTabs/UserTabs";
+import {ProfileData} from "@/pagesComponents";
 
 type Props = {
   params: {id: string, userName: string}
@@ -8,6 +8,6 @@ export default function Page (props: Props) {
   const {params: {id: userId, userName}} = props
 
   return (
-    <UserTabs userId={+userId}/>
+    <ProfileData userId={+userId}/>
   )
 }
