+++ src/pagesComponents/index.ts
@@ -1,0 +1,3 @@
+export { SignIn } from './signIn/ui/SignIn'
+export { ProfileData } from './profileData/ProfileData'
+export { UsersList } from './usersList/UsersList'
