+++ src/shared/lib/hooks/useDebounce.ts
@@ -1,3 +1,4 @@
+"use client"
 import { useEffect, useState } from 'react'
 
 export function useDebounce<T>(value: T, delay?: number): T {
