+++ src/shared/lib/index.ts
@@ -1,4 +1,7 @@
 export {useCustomToast} from './hooks/useCustomToast'
+export { useDebounce } from './hooks/useDebounce'
+export { useModal, ModalKey } from './hooks/useModal'
+export { useQueryParams } from './hooks/useQueryParams'
 export {MAIN_DOMAIN} from './constants/globalConstants'
 export { formatDate, combineFirstLastName } from './helpers/transformData'
 export { userTabsVariants } from './constants/userTabsVariants'
