+++ src/widget/index.ts
@@ -1,0 +1,5 @@
+export { ProfileHead } from './profileHead/ProfileHead'
+export { Header } from './header/Header'
+export { SideBar } from './sideBar/SideBar'
+export { UserTabsNavigation } from './userTabsNavigation/UserTabsNavigation'
+export { UserUploadedPhotosTab } from './userTabs/userUploadedPhotosTab/UserUploadedPhotosTab'
