+++ src/widget/profileHead/ProfileHead.module.scss
