+++ src/widget/userTabs/userUploadedPhotosTab/UserUploadedPhotosTab.module.scss
